Convert all colors to be generic over floats
Change all colors (Color, Rgb, Hsl, Hsv, Lab, Xyz) and traits (Mix and Saturation etc) to be
generic over floats.

This is a breaking change.

Since Rust defaults to f64, most of the color types will default to f64. For example Rgb::linear_rgb(1.0, 0.0, 0.0)
will be inferred as Rgb<f64>. To use f32, constants must be explicitly set like Rgb::linear_rgb(1.0_f32, 0.0, 0.0)
or the type explicitly annotated.

T::from( constant ).unwrap() etc was used in a number of places to convert constants to
num::Float trait ( via the num::Numcast trait).

For Hues, the Into<T> trait was not generic over floats due to unwrap() and possibility of
runtime panic and was separately implemented for f32 and f64 for Hue<32> and Hue<64>.
Also the Partial Eq trait could not be implemented for Hues as float and was removed.

The RgbPixel could not be made generic for [T;3] and (T,T,T) and was separately
implemented for f32, f64 and u8.

Using associated typed for Mix and Saturation was not ergonomic and the trait was
made generic over float.
s