Add the itu color spaces for rec.601 and rec.709
These share the same transfer function which is therefore a separate
struct instead of each sharing an implementation only. Strictly
speaking, the RgbSpace defined by Rec.709 could be shared with the sRGB
color space as it uses the exact same primaries and whitepoint but the
advantage due to the difference in transfer function seems negligible.