Provide tests and fix conversion
Positive and negative color differences were accidentally reversed. This
fixes that mistakes and adds full white and black as tests. For BT601
there are some reference values in a table that are also tested. The
tests vary on exact, ulp, and absolute difference based on the involved
calculation. Notably for BT709 the standard luminance weights differ
from exact Yxy colors for the primaries so that the results are as
expected only to some 12 bits (more than enough still).

This also introduces some additional trait implementations that were
either helpful or required for implementing the tests (comparison and
propagation of fundamental traits to their yuv standards).