Add ITU-Rec BT.2100 with HLG transfer function
This is for scene linear light, which does not depend on viewing
parameters in the transfer function. For PQ there are is no universal
reversible transfer function given, the recommendation only describes a
set of OETF and EOTF with which a reference viewing environment can
produce the proper luminance but this is not easily reversible.